+++ 1.8.9/src/main/java/io/github/axolotlclient/bridge/mixin/PlatformDispatchMixin.java
@@ -33,6 +33,7 @@ import io.github.axolotlclient.bridge.impl.AxoSpriteImpl;
 import io.github.axolotlclient.bridge.render.AxoSprite;
 import io.github.axolotlclient.mixin.MinecraftClientAccessor;
 import io.github.axolotlclient.modules.hypixel.autoboop.FilterListConfigurationScreen;
+import io.github.axolotlclient.modules.hypixel.bedwars.SessionStatsHudEntryConfigScreen;
 import io.github.axolotlclient.util.ThreadExecuter;
 import io.github.axolotlclient.util.Util;
 import net.minecraft.client.Minecraft;
@@ -146,5 +147,14 @@ public abstract class PlatformDispatchMixin {
 	@Overwrite
 	public static void autoBoop$openFiltersScreen(List<String> filters) {
 		Minecraft.getInstance().openScreen(new FilterListConfigurationScreen(filters, Minecraft.getInstance().screen));
+	}
+
+	/**
+	 * @author moehreag
+	 * @reason Implement bridge.
+	 */
+	@Overwrite
+	public static void bedwars$sessionstats$openEntryConfigScreen() {
+		Minecraft.getInstance().openScreen(new SessionStatsHudEntryConfigScreen(Minecraft.getInstance().screen));
 	}
 }